requirements.txt
# Core ML dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Chess engine and utilities
python-chess>=1.999

# Modal for distributed training
modal>=0.63.0

# Data processing
numpy>=1.24.0
tqdm>=4.65.0
datasets>=2.14.0  # For Hugging Face datasets

# ONNX export and inference
onnx>=1.14.0
onnxruntime>=1.15.0

# Misc utilities
requests>=2.31.0

